created method for create posts in thread, but it's not working

+++ internal/pkg/models/vote.go
@@ -3,7 +3,7 @@ package models
 //easyjson:json
 type Vote struct {
 	Nickname string `json:"nickname"`
-	Voice    int32  `json:"voice"`
+	Voice    int16  `json:"voice"`
 }
 
 //easyjson:json
